requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
pandas==2.2.3
scikit-learn==1.5.2
requests==2.32.3
pydantic==2.9.2
google-generativeai==0.8.3
httpx==0.27.2
